Add files via upload
Smart_Parking-master contains of the embedded codes that are uploaded to MCU which contains opening and closing of gates, sending IR Sensor data to our server, Verification of OTP's.

Smartpark-master contains our backend script on python (django) which stores the mobile no. of the user and also sends an OTP to the mobile no. It also shows the data of the vacant and the taken slots.